requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
docx==0.2.4
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
openai>=1.52.0
Flask
Flask-Cors
psycopg2-binary
pandas
numpy
SQLAlchemy
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.5
openai>=1.52.0
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
soupsieve==2.7
SQLAlchemy
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
